[UPDATE] s3 인터페이스와 구현체로 분리
